disposición de documentos y tratamiento de información

+++ src/main/java/com/example/demo/model/DOCUMENTO_PERSONA.java
@@ -20,6 +20,6 @@ public class DOCUMENTO_PERSONA {
     private Long id;
     private String descripcion;
     private int persona_id;
-    private int documento_id;
+    private Long documento_id;
     private int relacion_documento_id;
 }
